Fix bug in 'api/products' PUT method, refactor to return updated product.

+++ routes/api/product-routes.js
@@ -68,45 +68,54 @@ router.post('/', (req, res) => {
 });
 
 // update product
-router.put('/:id', (req, res) => {
-  // update product data
-  Product.update(req.body, {
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((product) => {
+router.put('/:id', async (req, res) => {
+  try {
+    // update product data
+    await Product.update(req.body, {
+      where: {
+        id: req.params.id,
+      }
+    });
+    // if the request body includes a tagIds field,
+    // need to update the product_tags join table.
+    if (req.body.tagIds) {
       // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
+      const productTags = await ProductTag.findAll({ 
+        where: { 
+          product_id: req.params.id 
+        } 
+      });  
       // get list of current tag_ids
       const productTagIds = productTags.map(({ tag_id }) => tag_id);
       // create filtered list of new tag_ids
       const newProductTags = req.body.tagIds
         .filter((tag_id) => !productTagIds.includes(tag_id))
         .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
+        return {
+          product_id: req.params.id,
+          tag_id,
+        };
+      });
+      console.log(newProductTags);
       // figure out which ones to remove
       const productTagsToRemove = productTags
         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
         .map(({ id }) => id);
-
       // run both actions
-      return Promise.all([
+      await Promise.all([
         ProductTag.destroy({ where: { id: productTagsToRemove } }),
         ProductTag.bulkCreate(newProductTags),
       ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
+    }
+    // get updated product and return it
+    const updatedProduct = await Product.findByPk(req.params.id, {
+      include: [{model: Category}, {model: Tag}]
     });
+    res.status(200).json(updatedProduct);
+    // res.status(200).json({message:'success'});
+  } catch (err) {
+    res.status(400).json(err);
+  }
 });
 
 router.delete('/:id', async (req, res) => {
